Edited the layout of mvc base models and updated app.js
In the new layout, the controller object receives all notifications
coming from the view or model.
When the model is updated it notifies the controller which notifies
the view to update.
If the view initiates a change to the model then it makes a round
trip, first notifying the controller who updates the model who
then notifies the controller it has been updated which then tells
the view to update as well.
This pattern ensures both model and view will always be in sync.